Elements: Finite Mixin Class
Add a mixin class for finite elements
