+++ src/particles/elements/Drift.H
@@ -12,6 +12,7 @@
 
 #include "particles/ImpactXParticleContainer.H"
 #include "mixin/regular.H"
+#include "mixin/finite.H"
 
 #include <AMReX_Extension.H>
 #include <AMReX_REAL.H>
@@ -22,7 +23,8 @@
 namespace impactx
 {
     struct Drift
-    : public elements::Regular<Drift>
+    : public elements::Regular<Drift>,
+      public elements::Finite
     {
         static constexpr auto name = "Drift";
         using PType = ImpactXParticleContainer::ParticleType;
@@ -33,7 +35,7 @@ namespace impactx
          * @param nslice number of slices used for the application of space charge
          */
         Drift( amrex::ParticleReal const ds, int const nslice )
-        : m_ds(ds), m_nslice(nslice)
+        : Finite(ds, nslice)
         {
         }
 
@@ -126,30 +128,6 @@ namespace impactx
             refpart.s = s + slice_ds;
 
         }
-
-        /** Number of slices used for the application of space charge
-         *
-         * @return positive integer
-         */
-        AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-        int nslice () const
-        {
-            return m_nslice;
-        }
-
-        /** Return the segment length
-         *
-         * @return value in meters
-         */
-        AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-        amrex::ParticleReal ds () const
-        {
-            return m_ds;
-        }
-
-    private:
-        amrex::ParticleReal m_ds; //! segment length in m
-        int m_nslice; //! number of slices used for the application of space charge
     };
 
 } // namespace impactx
