+++ src/particles/elements/Quad.H
@@ -12,6 +12,7 @@
 
 #include "particles/ImpactXParticleContainer.H"
 #include "mixin/regular.H"
+#include "mixin/regular.H"
 
 #include <AMReX_Extension.H>
 #include <AMReX_REAL.H>
@@ -22,7 +23,8 @@
 namespace impactx
 {
     struct Quad
-    : public elements::Regular<Quad>
+    : public elements::Regular<Quad>,
+      public elements::Finite
     {
         static constexpr auto name = "Quad";
         using PType = ImpactXParticleContainer::ParticleType;
@@ -38,7 +40,7 @@ namespace impactx
          */
         Quad( amrex::ParticleReal const ds, amrex::ParticleReal const k,
               int const nslice )
-        : m_ds(ds), m_k(k), m_nslice(nslice)
+        : Finite(ds, nslice), m_k(k)
         {
         }
 
@@ -148,30 +150,8 @@ namespace impactx
             refpart.s = s + slice_ds;
         }
 
-        /** Number of slices used for the application of space charge
-         *
-         * @return positive integer
-         */
-        AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-        int nslice () const
-        {
-            return m_nslice;
-        }
-
-        /** Return the segment length
-         *
-         * @return value in meters
-         */
-        AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-        amrex::ParticleReal ds () const
-        {
-            return m_ds;
-        }
-
     private:
-        amrex::ParticleReal m_ds; //! segment length in m
         amrex::ParticleReal m_k; //! quadrupole strength in 1/m
-        int m_nslice; //! number of slices used for the application of space charge
     };
 
 } // namespace impactx
