+++ src/particles/elements/Sbend.H
@@ -12,6 +12,7 @@
 
 #include "particles/ImpactXParticleContainer.H"
 #include "mixin/regular.H"
+#include "mixin/finite.H"
 
 #include <AMReX_Extension.H>
 #include <AMReX_REAL.H>
@@ -22,7 +23,8 @@
 namespace impactx
 {
     struct Sbend
-    : public elements::Regular<Sbend>
+    : public elements::Regular<Sbend>,
+      public elements::Finite
     {
         static constexpr auto name = "Sbend";
         using PType = ImpactXParticleContainer::ParticleType;
@@ -35,7 +37,7 @@ namespace impactx
          */
         Sbend( amrex::ParticleReal const ds, amrex::ParticleReal const rc,
                int const nslice)
-        : m_ds(ds), m_rc(rc), m_nslice(nslice)
+        : Finite(ds, nslice), m_rc(rc)
         {
         }
 
@@ -154,30 +156,8 @@ namespace impactx
 
         }
 
-        /** Number of slices used for the application of space charge
-         *
-         * @return positive integer
-         */
-        AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-        int nslice () const
-        {
-            return m_nslice;
-        }
-
-        /** Return the segment length
-         *
-         * @return value in meters
-         */
-        AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
-        amrex::ParticleReal ds () const
-        {
-            return m_ds;
-        }
-
     private:
-        amrex::ParticleReal m_ds; //! segment length in m
         amrex::ParticleReal m_rc; //! bend radius in m
-        int m_nslice; //! number of slices used for the application of space charge
     };
 
 } // namespace impactx
